% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Beta.R
\name{alphaBeta_to_meanVar}
\alias{alphaBeta_to_meanVar}
\title{fit a 2-component beta mixture}
\usage{
alphaBeta_to_meanVar(alpha, beta)
}
\arguments{
\item{y}{ration values between 0 and 1}

\item{inits}{Initial value to fit the mixture model. A vector with elements alpha1, beta1, alpha2, beta2 and pi1.}

\item{model}{Character specifying E or V model. E model fits the mixture model with equal variance while V model doesn't put any constraint.}
}
\value{
A vector consisting parameter estimates of mu1, mu2, phi1, phi2, pi1, logLik and BIC. For 0-inflated model, mu1=phi1=0.
}
\description{
can impose the two component to have equal variance (model='E') or unequal variance (model='V')
}

