\name{demo.siber.data}
\alias{demo.siber.data}
\docType{data}
\title{
Some SIBER demo data
}
\description{
Some SIBER demo data
}
\usage{data(demo.siber.data)}
\format{
  A data frame with 180 observations and 4 variables
  \describe{
    \item{\code{iso1}}{Isotope 1}
    \item{\code{iso2}}{Isotope 2}
    \item{\code{group}}{Group number}
    \item{\code{community}}{Community number}
  }
}
\details{
The is some demo data used by the SIBER package
}
\keyword{datasets}
