% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoidTransform.R
\name{ellipsoidTransform}
\alias{ellipsoidTransform}
\title{Apply a normalisation transformation to vectors of data onto ellipsoids}
\usage{
ellipsoidTransform(x, SigSqrt, mu)
}
\arguments{
\item{x}{the vector of data points to be transformed}

\item{SigSqrt}{the inverse of the covariance matrix}

\item{mu}{the vector of means of the ellipse}
}
\value{
A vector of transformed data points
}
\description{
Takes a vector \code{x} and transforms the points onto the same geometry of 
a normalised ellipse given by the inverse of the covariance matrix 
\code{SigSqrt} and the location \code{mu}
}
