% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.posterior.means.R
\name{extract.posterior.means}
\alias{extract.posterior.means}
\title{Extract posterior means from call to \code{\link{siber.MVN}}}
\usage{
extract.posterior.means(siber, post)
}
\arguments{
\item{siber}{a siber object as created by \code{\link{create.siber.object}}}

\item{post}{a list containing the posterior estimated parameters fitted to
each group within every community. See \code{\link{siber.MVN}} which creates
this object for details.}
}
\value{
A list of length n.communities with each entry representing a
\code{n.draws * 2 * n.groups} array of rows equal to the number of posterior
samples, 2 columns representing the two means of the multivariate data and
n.groups the number of groups within the focal community.
}
\description{
This function extracts the posterior means from a call to
\code{\link{siber.MVN}} which can then be used to calculate bayesian layman
metrics. This function is designed to create an array of posterior means
that is more easily interrogated for plotting and summary statistics.
}

