% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siberRaw2.R
\name{fitSIBERonMat}
\alias{fitSIBERonMat}
\title{fit SIBER on a matrix, one row at a time}
\usage{
fitSIBERonMat(mat, model = "NL", prune = TRUE, q = 0.03,
  parallel = TRUE, core = 10)
}
\arguments{
\item{mat}{matrix of expression, methylation and so on}

\item{model}{model as in SIBER}

\item{prune}{logical, whether to prune sigma1, sigma2 for with a specified percentile; this is to obtain stable BI estimate due to extremely small
sigma estimate, especially in the V model}

\item{parallel}{logical}

\item{core}{number of cores to register for parallel computing}
}
\value{
a matrix
}
\description{
this function enables parallel through plyr package
}
\references{
Tong, P., Chen, Y., Su, X. and Coombes, K. R. (2012). Systematic Identification of Bimodally Expressed Genes Using RNAseq Data. Bioinformatics, submitted.
}

