% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\docType{data}
\name{geese2demo}
\alias{geese2demo}
\title{A single group of the geese data}
\format{
A 3 column, 251 row matrix containing the plasma data for the 8
  groups of gees as consumers. Columns are in the order Group which is an
  integer that determines which of the 8 groups the observation belons. The
  second and third columns are d13C and d15N values derived from the blood
  plasma for each observation. Retained here as legacy from now defunct
  package siar. Note that the order of the isotope data has been swapped
  since siar in order to present d13C as the first isotope and hence on the
  x-axis by default.
}
\usage{
data(geese2demo)
}
\description{
A dataset for a single group of geese (as consumers) for two isotope tracers.
Intended for use in a Stable Isotope Mixing Model.
}
\author{
Rich Inger
}
\keyword{datasets}
