% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen.circle.R
\name{gen.circle}
\alias{gen.circle}
\title{Create a sequence of points on a circle}
\usage{
gen.circle(n = 100, r)
}
\arguments{
\item{n}{the number of points to create around the circle. Defaults to 100.}

\item{r}{the radius of the circle to create.}
}
\value{
A 2 x n matrix of x and y coordinates of points on a circle.
}
\description{
This is a helper function that creates a sequence of points on a circle of
radius \code{r} as a resolution determined by \code{n}. It is not intended
for direct calling, and is used by the ellipse plotting function
\code{\link{add.ellipse}}.
}
\examples{
plot(gen.circle(n = 20, r = 1), asp = 1, pch = 19, cex = 0.6)
}

