% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCircle.R
\name{genCircle}
\alias{genCircle}
\title{Create a sequence of points on a circle}
\usage{
genCircle(n = 100, r)
}
\arguments{
\item{n}{the number of points to create around the circle. Defaults to 100.}

\item{r}{the radius of the circle to create.}
}
\value{
A 2 x n matrix of x and y coordinates of points on a circle.
}
\description{
This is a helper function that creates a sequence of points on a circle of 
radius \code{r} as a resolution determined by \code{n}. It is not intended 
for direct calling, and is used by the ellipse plotting function 
\code{\link{addEllipse}}. NB not an exported function.
}
