% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hullArea.R
\name{hullArea}
\alias{hullArea}
\title{Calculate the area of a convex hull given its coordinates}
\usage{
hullArea(x, y)
}
\arguments{
\item{x}{a vector of x-axis data}

\item{y}{a vector of y-axis data}
}
\value{
a scalar representing the area of the convex hull in units of 
\code{x * y}; i.e. most commonly in permille squared for isotope data.
}
\description{
Given the coordinates of a convex hull (i.e. a polygon), this function
calculates its area. Not intended for direct use outside of 
\code{\link{siberConvexhull}}.
}

