% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kapow.R
\name{kapow}
\alias{kapow}
\title{KAPOW!}
\usage{
kapow(cd = 7, ng = 25, n = 50, sc = 10, do.plot = TRUE)
}
\arguments{
\item{cd}{sets the random seed to this}

\item{ng}{the number of ellipses to draw}

\item{n}{the number of data points to simulate per group, but never displayed}

\item{sc}{the scaling factor the rwishart sigma called by
\code{\link[stats:rWishart]{stats::rWishart()}}}

\item{do.plot}{a logical indicating whether the plot should be printed
(defaults to TRUE).}
}
\value{
A ggplot object
}
\description{
This function packs a punch and makes a pretty figure.
}
