% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Beta.R
\name{meanVar_to_alphaBeta}
\alias{meanVar_to_alphaBeta}
\title{convert mean/variance parameterization to Beta(alpha, beta) parameterization for a beta distribution}
\usage{
meanVar_to_alphaBeta(m, v)
}
\description{
convert mean/variance parameterization to Beta(alpha, beta) parameterization for a beta distribution
}

