% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\docType{data}
\name{mongoose}
\alias{mongoose}
\title{A set of isotope observations for mongooses nested within packs}
\format{
A 4 column, 783 row data.frame object containing unique individual 
mongoose identifiers in the first column "indiv.id"; an integer identifier for 
the pack to which each individual belongs in "pack"; Delta 13 Carbon values 
"c13; and Delta 15 Nitrogen values in "n15". See the paper Sheppard et al 
2018 https://doi.org/10.1111/ele.12933 for more details, although N.B. 
the data here are provided as an example, not as a reproducible analysis of 
that paper.
}
\usage{
data(mongoose)
}
\description{
A dataset of multiple isotopes per individual mongooses nested within packs 
where the goal is to understand isotopic niche occupancy of individuals 
with respect to their own pack.
}
\author{
Harry Marshall
}
\keyword{datasets}
