% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCentroidVectors.R
\name{plotCentroidVectors}
\alias{plotCentroidVectors}
\title{Plot the pairwise distances and angles describing the difference between 
centroids}
\usage{
plotCentroidVectors(centroids, upper = TRUE)
}
\arguments{
\item{centroids}{the list containing distance and angle matrices as returned 
by \code{\link{siberCentroids}}.}

\item{upper}{a logical determining whether to plot the upper or lower 
triangle of angles. DEfaults to TRUE which is the upper triangle and 
returns the angle from the second ellipse to the first by centering on the 
first centroid.}
}
\value{
A nice plot. You can get the corresponding matrices used to generate
  the plots if you ask for it nicely: thedata <-
  plotCentroidVectors(centroids)
}
\description{
Plots the posterior densities
}
