% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGroupEllipses.R
\name{plotGroupEllipses}
\alias{plotGroupEllipses}
\title{Adds ellipses to an existing plot for each of your groups}
\usage{
plotGroupEllipses(siber, plot.args = list(), iso.order = c(1, 2), ...)
}
\arguments{
\item{siber}{a siber object as created by createSiberObject}

\item{plot.args}{a list of plotting arguments for passing to
\code{\link{addEllipse}}. See \code{\link{addEllipse}} for details of the
options, and you can also pass additional arguments such as line widths and
styles. See also the demonstration scripts for examples of use.}

\item{iso.order}{a vector of length 2, either \code{c(1,2)} or \code{c(2,1)}.
The order determines which of the columns of raw data are plotted on the x
(1) or y (2) axis. N.B. this will be deprecated in a future release, and
plotting order will be acheived at point of data-entry.}

\item{...}{additional arguments to be passed to \code{\link{addEllipse}}.}
}
\value{
Ellipses, drawn as lines on an existing figure.
}
\description{
This function loops over each community and group within, and plots an
ellipse around the data. See demonstration scipts for more examples.
}

