% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEA.posterior.R
\name{SEA.posterior}
\alias{SEA.posterior}
\title{Calculate the SEA based on a posterior distribution of Sigma}
\usage{
SEA.posterior(post)
}
\arguments{
\item{post}{a matrix of postior covariance matrix and mean estimates for a
bivariate ellipse. In SIBER, this is typically one list element of the
object returned by \code{link{siber.MVN}}.}
}
\value{
A vector of posterior Bayesian Standard Ellipse Areas (SEA_B)
}
\description{
This function loops over each posterior draw of a single group's
Bayesian bivariate ellipse and calculates the Standard Ellipse Area (SEA)
for each draw, thereby generating a distribution of SEA esimates. Not
intended for direct calling outside of \code{\link{siber.ellipses}}.
}

