% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBER-package.R
\docType{package}
\name{SIBER-package}
\alias{SIBER}
\alias{SIBER-package}
\title{SIBER: Stable Isotope Bayesian Ellipses in R}
\description{
Fits bi-variate ellipses to stable isotope data using Bayesian inference with the aim being to describe and compare their isotopic niche.
}
\author{
\strong{Maintainer}: Andrew Jackson \email{jacksoan@tcd.ie} (\href{https://orcid.org/0000-0001-7334-0434}{ORCID})

Authors:
\itemize{
  \item Andrew Parnell (\href{https://orcid.org/0000-0001-7956-7939}{ORCID})
}

}
\keyword{internal}
