% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siber.convexhull.R
\name{siber.convexhull}
\alias{siber.convexhull}
\title{Calculate metrics and plotting information for convex hulls}
\usage{
siber.convexhull(x, y)
}
\arguments{
\item{x}{a vector of x-axis data}

\item{y}{a vector of y-axis data}
}
\value{
A list of length four comprising:
 \itemize{
   \item{TA}{the area of the convex hull.}
   \item{hullX}{the x-coordinates of the points describing the convex hull.}
   \item{hullY}{the y-coordinates of the points describing the convex hull.}
   \item{ind}{the indices of the original data in \code{x} and \code{y} that
               form the boundaries of the convex hull.}

 }
}
\description{
This function calculates the area of the convex hull describing a set of
bivariate points, and returns other information useful for plotting the hull.
}
\examples{
x <- rnorm(15)
y <- rnorm(15)
siber.convexhull(x, y)
}

