% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siber.ellipses.R
\name{siber.ellipses}
\alias{siber.ellipses}
\title{Calculate the point estimates of the Layman metrics for each community}
\usage{
siber.ellipses(corrected.posteriors)
}
\arguments{
\item{siber}{a siber object as created by create.siber.object.R}
}
\value{
A 6 x m matrix of the 6 Layman metrics of dX_range, dY_range, TA,
CD, MNND and SDNND in rows, for each community by column
}
\description{
This function loops over each community, determines the centre of mass
(centroid) of each of the groups comprising the community using the basic
\code{\link{mean}} function independently on the marginal x and y vectors,
and calculates the corresponding 6 Layman metrics based on these points.
}
\examples{
data(demo.siber.data)
my.siber.data <- create.siber.object(demo.siber.data)
community.metrics.ML(my.siber.data)
}

