% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siberEllipses.R
\name{siberEllipses}
\alias{siberEllipses}
\title{Calculate the point estimates of the Layman metrics for each community}
\usage{
siberEllipses(corrected.posteriors)
}
\arguments{
\item{corrected.posteriors}{the Bayesian ellipses as fitted to z-score 
transformed data returned by \code{\link{siberMVN}}.}
}
\value{
A matrix of with each column representing a unique community.group 
combination, and each row an independent posterior estimate of the 
Standard Ellipse Area SEA_B.
}
\description{
This function loops over each community, determines the centre of mass 
(centroid) of each of the groups comprising the community using the basic 
\code{\link[base]{mean}} function independently on the marginal x and y vectors,
and calculates the corresponding 6 Layman metrics based on these points.
}
