% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siberRaw2.R
\name{updateBImat}
\alias{updateBImat}
\title{update the result from BI data frame.}
\usage{
updateBImat(dat)
}
\arguments{
\item{dat}{a data frame with results fit from SIBER}
}
\description{
update the result from BI data frame.
}

