% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siberRaw2.R
\name{fitNL}
\alias{fitNL}
\title{Fit Normal Mixture Model
The function fits a two-component normal mixture model.}
\usage{
fitNL(y, d = NULL, model = "E")
}
\arguments{
\item{y}{A vector representing the RNAseq raw count.}

\item{d}{A vector of the same length as y representing the normalization constant to be applied to the data. For the LN model, the original data would be devided by this vector.}

\item{model}{Character specifying E or V model. E model fits the mixture model with equal variance while V model doesn't put any constraint.}
}
\value{
A vector consisting parameter estimates of mu1, mu2, sigma1, sigma2, pi1, logLik and BIC.
}
\description{
The parameter estimates from log normal mixture is obtained by taking logarithm and fit normal mixture. We use
mclust package to obtain parameter estimates of normal mixture model. In particular, \eqn{log_{base}(\frac{y+eps}{d})} is used to
fit to normal mixture model.
}
\details{
With this function, three models can be fitted: (1) log normal mixture with equal variance (E model);
(2) Generalized Poisson mixture with unequal variance (V model); (3) 0-inflated log normal model.
The 0-inflated log normal has the following density function:

\eqn{P(Y=y)=\pi D(y) + (1-\pi)LN(\mu, \sigma)} where D is the point mass at 0 while \eqn{LN(\mu, \sigma)} is the density
of log normal distribution with mean \eqn{\mu} and variance \eqn{\sigma^2}.

The rule to fit 0-inflated model is that the observed percentage of count exceeds the user specified threshold. This
rule overrides the model argument (E or V) when observed percentae of zero count exceeds the threshold.
}
\references{
Tong, P., Chen, Y., Su, X. and Coombes, K. R. (2012). Systematic Identification of Bimodally Expressed Genes Using RNAseq Data. Bioinformatics, submitted.
}

