% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siber.ellipses.R
\name{siber.ellipses}
\alias{siber.ellipses}
\title{Calculate the point estimates of the Layman metrics for each community}
\usage{
siber.ellipses(corrected.posteriors)
}
\arguments{
\item{corrected.posteriors}{the Bayesian ellipses as fitted to z-score
transformed data returned by \code{\link{siber.MVN}}.}
}
\value{
A matrix of with each column representing a unique community.group
combination, and each row an independent posterior estimate of the
Standard Ellipse Area SEA_B.
}
\description{
This function loops over each community, determines the centre of mass
(centroid) of each of the groups comprising the community using the basic
\code{\link[base]{mean}} function independently on the marginal x and y vectors,
and calculates the corresponding 6 Layman metrics based on these points.
}
\examples{
data(demo.siber.data)
my.siber.data <- create.siber.object(demo.siber.data)
community.metrics.ML(my.siber.data)
}

