% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBER.R
\docType{package}
\name{SIBER}
\alias{SIBER}
\alias{SIBER-package}
\title{SIBER: A package for fitting Bayesian Ellipses to Stable Isotope Data}
\description{
The SIBER package provides tools for fitting multivariate normal 
distributions to bivariate data using Bayesian Inference. These
distributions can then be used to calculate probability 
distributions of Standard Ellise Areas for comparing groups of data,
or to calculate the 6 Layman metrics for comparing entire communities.
}
\section{SIBER functions}{

The SIBER functions ...
}

