% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siberEllipses.R
\name{siberEllipses}
\alias{siberEllipses}
\title{Calculate the Bayesian Standard Ellipse Area for all groups}
\usage{
siberEllipses(corrected.posteriors)
}
\arguments{
\item{corrected.posteriors}{the Bayesian ellipses as returned by
\code{\link[=siberMVN]{siberMVN()}}.}
}
\value{
A matrix of with each column containing the posterior estimates of
the SEA.
}
\description{
This function loops over each group within each community and calculates the
posterior distribution describing the corresponding Standard Ellipse Area.
}
